renamed file names, and introduced DefaultFormFooter
