+++ Client/src/Components/ResearchPaperAnalysis/ResearchPaperAnalysisView.tsx
@@ -1,7 +1,7 @@
 import { Box, Typography, useTheme } from '@material-ui/core'
 import React, { useState } from 'react'
 
-import { DatasetView } from '../DatasetUpload/DatasetUploadView'
+import { DatasetView } from '../DatasetUpload/DatasetView'
 import { FileUploadForm } from '../FileUpload/FileUploadForm'
 import { IDatasetModel } from '../../Models/Datasets/IDatasetModel'
 import Loader from "react-loader-spinner"
